Make reference testing base class.
