+++ unittests/test_mpcpy.py
@@ -1,0 +1,50 @@
+# -*- coding: utf-8 -*-
+"""
+This module contains the classes for testing in mpcpy.
+
+"""
+
+from abc import ABCMeta
+import unittest
+from mpcpy import utility
+import pandas as pd
+from pandas.util.testing import assert_frame_equal
+import os
+
+class test_mpcpy(unittest.TestCase):
+    '''General test methods for testing in mpcpy.
+    
+    '''
+    
+    __metaclass__ = ABCMeta;
+    
+    def get_ref_path(self):
+        ref_path = utility.get_MPCPy_path() + 'unittests/references/' + self.__module__.split('.')[-1] + '/' + self.__class__.__name__;
+        
+        return ref_path;
+        
+    def check_df_timeseries(self, df_test, ref_file_name):
+        # Define reference file
+        ref_file_path = self.get_ref_path() + '/' + ref_file_name;
+        try:
+            df_ref = pd.read_csv(ref_file_path, index_col='Time');
+            df_ref.index = pd.to_datetime(df_ref.index);
+            df_ref = df_ref.tz_localize('UTC')
+            assert_frame_equal(df_test, df_ref);
+        except IOError:
+            ref_file_dir = self.get_ref_path();
+            if not os.path.exists(ref_file_dir):
+                os.makedirs(ref_file_dir);
+            df_test.to_csv(ref_file_path);
+            
+    def check_df_parameter(self, df_test, ref_file_name):
+        # Define reference file
+        ref_file_path = self.get_ref_path() + '/' + ref_file_name;
+        try:
+            df_ref = pd.read_csv(ref_file_path, index_col='Name');
+            assert_frame_equal(df_test, df_ref, check_dtype=False);
+        except IOError:
+            ref_file_dir = self.get_ref_path();
+            if not os.path.exists(ref_file_dir):
+                os.makedirs(ref_file_dir);
+            df_test.to_csv(ref_file_path);
